+++ datasophon-common/src/main/java/com/datasophon/common/vo/ClusterServiceInstanceRoleGroupVO.java
@@ -30,7 +30,7 @@ import java.io.Serializable;
 public record ClusterServiceInstanceRoleGroupVO(
         Long id,
         String roleGroupName,
-        Integer serviceInstanceId,
+        Long serviceInstanceId,
         String serviceName,
         Long clusterId,
         String roleGroupType,
