+++ datasophon-service/src/main/java/com/datasophon/api/strategy/HBaseHandlerStrategy.java
@@ -73,7 +73,7 @@ public class HBaseHandlerStrategy extends ServiceHandlerAbstract implements Serv
      */
     @Override
     protected ConnectionInfo.ConnectionInfoBuilder getServiceSpecificConnectionInfo(
-            Long clusterId, Integer serviceInstanceId, Map<String, String> configMap) {
+            Long clusterId, Long serviceInstanceId, Map<String, String> configMap) {
         try {
             logger.info("开始获取HBase服务连接信息，集群ID: {}, 服务实例ID: {}", clusterId, serviceInstanceId);
