+++ datasophon-service/src/main/java/com/datasophon/api/strategy/RedisHandlerStrategy.java
@@ -108,7 +108,7 @@ public class RedisHandlerStrategy extends ServiceHandlerAbstract implements Serv
          */
         @Override
         protected ConnectionInfo.ConnectionInfoBuilder getServiceSpecificConnectionInfo(
-                        Long clusterId, Integer serviceInstanceId, Map<String, String> configMap) {
+                        Long clusterId, Long serviceInstanceId, Map<String, String> configMap) {
                 try {
                         logger.info("开始获取Redis服务连接信息，集群ID: {}, 服务实例ID: {}", clusterId, serviceInstanceId);
